+++ NEWS.md
@@ -1,6 +1,7 @@
 # ispv (development version)
 
 * added function for loading country-wide ISCO-based data
+* added function for loading country-wide NACE-based data
 
 # ispv 0.1.0